[Bug Fix] Solve catkin_make problems
Previously, catkin_make wasn't able to work properly since some CMakeLists.txt files were not correct, so it was possible to build the workspace only with the build and devel folder already created when the workspace was built using also files not present in the origin repository.

There are still problems with dependencies, so usually it is necessary to build the workspace with catkin_make twice, since the first time some errors will be raised, but then the second time it should work. This will be solved in future updates.