Commit before major future changes
Since there is a problem with the topic names due to the fact that some nodes rely on the automatic topic names remapping caused by the group tag in launch files and some other don't, major changes to the code are in order. Hence, a safety commit is done now to have on the remote repository the last working version of the code.