Test broadcasting for DS
To test how a broadcast message is sent with the adhoc_communication paclage, robots send a broadcast message when it has selected a new optimal docking station. The message is successfully received by the other robots.

In future updates the message will be sent not when a new optimal docking station is selected but when a previously undiscovered docking station is found.