Improve auction management
Now when a robot receives an auction, it checks if the auctioned docking station is the one the robot is interested in. If no, it ignores the auction (and its result); if yes, it sends its bid and when the auction is finished it checks if he can go to recharge or if it has to go in queue and wait.

Notice that at the moment the robots do not consider their current state (i.e., if they need to recharge or not) when they receive an auction.