Improve auctions system
Now a robot can go to a dockign station, periodically perform an auction to check if now it can go the docking station for which it was in queue, leave a docking station in case another robot wins an auction, etc.

There is still a problem when a robot accesses a docking station, because if a robot was previously occupying it, that robot needs some time to leave the docking station: to be more precise, it needs some time to compute the next frontier to be visited, and meanwhile it keeps the docking station busy. This problem requires a better management of the signal to inform the other robots that a docking station is now vacant.