Improve check for vacancy of target DS
Previously, if some bid messages were lost, it was possible that a robot tried to occupy its target docking station even if it was already occupied by another robot. Now, before occupying a docking station, the robot stops to check the vacancy of the docking station.

For the moment the check does not produce any result (the robot simply stops forever, waiting for a result).