Improve callback management
Now the callback functions are handled in a more clean way: instead of letting them directly modify the state of the robot, they are used to store the next state of the robot; later, when it is safe to update the state of the robot, the change is made.

Notice that there is still a problem due to the fact that sometimes, for the moment, it is possible for instance to change the state of the robot from "charging" to "going_charging", which is clearly nonsense.