Add "lock" for the update of the currently optimal DS
The robot can compute the currently optimal docking station anytime, but can update it only if it is not participating to an auction: this is done to avoid that it wins the access to a certain docking station but then goes recharging to another one. Probably there are still some other checks to be done, to be sure that this scenario cannot really happen.