Improve charging likelihood management
Now all the functions that compute a part of the charging likelihood function should work properly. For the moment they still use the euclidean distance where they have to compute a distance, but there is already implemented a function that uses the Dijsktra algorithm, it's just a matter of calling this function where it is needed: it will be done in future updates.