Temporary bug fix
Now the bug caused by the strange segmentation fault seems not to happen anymore. Notice however that there are part of codes (in particular the ones related to the computation of the MST for the Ds graph) that are commented out, so it may be that they are the ones responsible for the bug.